+++ packages/core/src/type/type.ts
@@ -1,3 +1,12 @@
-export type VersionType = "git_commit_hash" | "package_version" | "custom_version";
-
+/*
+ * @Author: yinbing.liu
+ * @Date: 2023-08-08 15:57:12
+ * @LastEditors: yinbing.liu
+ * @LastEditTime: 2023-08-16 17:59:29
+ * @FilePath: /cache-updat/packages/core/src/type/type.ts
+ * @Description: core types
+ */
+export type VersionTypeField = "git_commit_hash" |
+    "package_version" |
+    "custom_version";