+++ packages/core/src/util/enum.ts
@@ -1,0 +1,12 @@
+/*
+ * @Author: yinbing.liu
+ * @Date: 2023-08-14 15:50:09
+ * @LastEditors: yinbing.liu
+ * @LastEditTime: 2023-08-16 18:00:04
+ * @FilePath: /cache-updat/packages/core/src/util/enum.ts
+ * @Description: 
+ */
+
+
+export const CACHED_UPDATE_VERSION = "cache-update-version_";
+export const JSON_FILE_NAME = 'version-web-update-notify'