+++ packages/vitePlugin/src/types/type.ts
@@ -1,0 +1,11 @@
+
+
+
+
+import type {Options} from "@plugin-web-cache-update/core"
+
+
+
+export type PluginsOptions = Options & {
+    indexHtmlFilePath?: string,
+}