+++ packages/vitePlugin/CHANGELOG.md
@@ -1,5 +1,72 @@
 # @plugin-web-cache-update/vite-plugin
 
+## 1.2.5
+
+### Patch Changes
+
+- fix: mes
+
+## 1.2.4
+
+### Patch Changes
+
+- sdfsd
+
+## 1.2.3
+
+### Patch Changes
+
+- sdfsdf
+
+## 1.2.2
+
+### Patch Changes
+
+- fix: messgae
+- Updated dependencies
+  - @plugin-web-cache-update/core@1.1.2
+
+## 1.2.1
+
+### Patch Changes
+
+- fix: core
+- Updated dependencies
+  - @plugin-web-cache-update/core@1.1.1
+
+## 1.2.0
+
+### Minor Changes
+
+- fix:core ....
+- fix: unbuild
+- fix: core unbuild
+
+### Patch Changes
+
+- fix: unbuild config
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+  - @plugin-web-cache-update/core@1.1.0
+
+## 1.1.3
+
+### Patch Changes
+
+- fix:core
+- Updated dependencies
+  - @plugin-web-cache-update/core@1.0.2
+
+## 1.1.2
+
+### Patch Changes
+
+- fix: core create
+- Updated dependencies
+  - @plugin-web-cache-update/core@1.0.1
+
 ## 1.1.1
 
 ### Patch Changes